requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
Js2Py==0.74
pycparser==2.21
pyjsparser==2.7.1
PySocks==1.7.1
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
